This commit will combine useEffect call to set the username and status
If the user checks out a server, then the new value will be Check In. Once its checked out the useEffect call will push the currently logged in User's username to the server db in the user: property. And once checked in, it will make the user property in server collection to an empty string.