This commit will fix the not available status for checkout system
Added  conditions in place to check if the user can checkout a server and perform actions based on if the status is available or if someone was able to checkout the server before the request. This is to avoid conflict when two users logged into the system try to checkout a server at the same time.